Reorganize file structure to match project
