+++ client/src/pages/Profile/Profile.css
@@ -1,0 +1,29 @@
+.sky {
+  background-image: url(../../components/Images/sky1.png);
+}
+
+.userImg {
+  display: flex;
+  justify-content: center;
+  margin: auto;
+}
+
+.photo {
+  width: 160px;
+  height: 160px;
+  border-radius: 80px;
+}
+
+h1,
+h4,
+h5 {
+  color: white;
+  text-align: center;
+}
+
+li {
+  list-style-type: none;
+  display: inline;
+  padding: 40px;
+  color: white;
+}
